+++ app/list_versions.rb
@@ -2,7 +2,7 @@
 
 =begin
 $META_START$
-name: list-version
+name: list-versions
 summary: Lists application versions.
 basic: false
 class: PatchKitTools::ListVersionsTool
@@ -11,48 +11,45 @@ $META_END$
 
 require_relative 'core/patchkit_api.rb'
 require_relative 'core/patchkit_tools.rb'
+require_relative 'core/base_tool2.rb'
 
 module PatchKitTools
-  class ListVersionsTool < PatchKitTools::BaseTool
+  class ListVersionsTool < PatchKitTools::BaseTool2
     attr_reader :versions_list
 
     SORT_MODES = ["desc", "asc"]
 
-    def initialize
-      super("list-versions",
-            "Lists application versions.",
-            "-s <secret> [optional]")
-
-      self.display_limit = -1
-      self.sort_mode = "desc"
+    def initialize(argv = ARGV)
+      super(argv, "list-versions", "Lists application versions.", "-s <secret> [optional]")
+      @display_limit = -1
+      @sort_mode = "desc"
+      @format = :yaml
     end
 
     def parse_options
       super do |opts|
         opts.separator "Mandatory"
 
-        opts.on("-s", "--secret <secret>",
-          "application secret") do |secret|
-          self.secret = secret
-        end
+        opts.on("-s", "--secret <secret>", "application secret") { |v| @secret = v }
 
         opts.separator ""
-
         opts.separator "Optional"
 
         opts.on("-a", "--apikey <apikey>",
-          "user API key (when supplied draft version is also listed)") do |api_key|
-          self.api_key = api_key
-        end
+                "user API key (when supplied draft version is also listed)") { |v| @api_key = v }
 
         opts.on("-l", "--displaylimit <display_limit>", Integer,
-          "limit of displayed versions; -1 = infinite (default: #{self.display_limit})") do |display_limit|
-            self.display_limit = display_limit
+                "limit of displayed versions; -1 = infinite (default: #{@display_limit})") do |v|
+          @display_limit = v
         end
 
         opts.on("-m", "--sortmode <sort_mode>",
-          "sort mode; #{SORT_MODES.join(", ")} (default: #{self.sort_mode})") do |sort_mode|
-            self.sort_mode = sort_mode
+                "sort mode; #{SORT_MODES.join(", ")} (default: #{@sort_mode})") do |v|
+          @sort_mode = v
+        end
+
+        opts.on('-f', '--format <format>', 'output formtat (default: yaml)') do |v|
+          @format = v.to_sym
         end
       end
     end
@@ -61,25 +58,32 @@ module PatchKitTools
       check_if_option_exists("secret")
       check_if_valid_option_value("sort_mode", SORT_MODES)
 
-      resource_name = "1/apps/#{self.secret}/versions"
-      resource_name << "?api_key=#{self.api_key}" unless self.api_key.nil?
+      resource_name = "1/apps/#{@secret}/versions"
+      resource_name << "?api_key=#{@api_key}" unless @api_key.nil?
 
       results = PatchKitAPI.get(resource_name)
-      results = results.sort_by {|version| self.sort_mode == "asc" ? version[:id] : -version[:id]}
+      results = results.sort_by {|version| @sort_mode == "asc" ? version[:id] : -version[:id]}
 
       @versions_list = results
 
-      if(self.display_limit > -1)
-        results = results[0,[self.display_limit, results.length].min]
-      end
+      results = results[0, [@display_limit, results.length].min] if @display_limit > -1
 
-      results.each do |version|
-        puts "|-- #{version[:label]} (#{version[:id]})"
+      puts case @format
+             when :yaml
+               YAML.dump(results)
+             when :json
+               JSON.generate(results)
+             else
+               raise "unknown format: #{@format}"
+             end
 
-        version.each do |key, value|
-          puts "|   |-- #{key}: #{value}"
-        end
-      end
+      # results.each do |version|
+      #   puts "|-- #{version[:label]} (#{version[:id]})"
+
+      #   version.each do |key, value|
+      #     puts "|   |-- #{key}: #{value}"
+      #   end
+      # end
     end
   end
 end