+++ CHANGELOG.md
@@ -4,15 +4,36 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0]
+### Added
+- channel-link-version command
+- channel-make-version command
+
+### Changed
+- All `--apikey` parameters renamed to `--api-key`
+- diff-version parameter `--diffsummary` renamed to `--out-diff-summary-file`
+- list-versions parameter `--displaylimit` renamed to `--display-limit`
+- list-versions parameter `--sortmode` renamed to `--sort-mode`
+- update-version parameter `--sortmode` renamed to `--sort-mode`
+- update-version parameter `--changelogfile` renamed to `--changelog-file`
+- make-version parameter `--changelogfile` renamed to `--changelog-file`
+- make-version parameters `-x` and `--override-draft` does not require second parameter
+- make-version parameter `--publish` does not require second parameter
+- upload-version parameter `--diffsummary` renamed to `--diff-summary-file`
+- upload-version parameter `--waitforjob` renamed to `--wait-for-job`
+- publish-version parameter `--wait-until-published` renamed to `--wait`
+- Progress bar: Limited refresh frequency to 2 per second. Makes the stdout stream smaller (good for logging).
+- list-version command renamed to list-versions
+- list-versions generating yaml or json (depending on -f param)
+- Refactoring of model abstraction layer
+- patchkit_api http errors are now APIErrors (without stacktrace)
+
 ## [2.5.0]
 ### Added
 - make-version --import-app-secret parameter
 - make-version --import-version parameter
 - make-version --import-copy-label parameter
 - make-version --import-copy-changelog parameter
-
-### Changed
-- Refactoring of model abstraction layer
 
 ## [2.4.0]
 ### Added